Use Abstract Method for the validator
